corrected template use

+++ trunk/ghmm/ghmm++/sequence_test.cpp
@@ -2,7 +2,7 @@
 
 int main()
 {
-  XMLIO_Array<int> my_array;
+  XMLIO_Array<double> my_array;
   my_array.XMLIO_getCharacters("1 1 1 1");
   my_array.push_back(1);
   my_array.print();
